grsec: Apply grsecurity-3.0-3.13.2-201402111747.patch
commit 596b24936ed3687455327c3d26a8a820263a1f88
Author: Brad Spengler <spender@grsecurity.net>
Date:   Tue Feb 11 17:33:49 2014 -0500

    [PATCH] random: fix overflow for big nbits values in credit_entropy_bits()

    Commit 30e37ec516ae "random: account for entropy loss due to overwrites"
    introduced an overflow in the arithmetics of credit_entropy_bits() when
    CONFIG_GRKERNSEC_RANDNET is enabled as the latter quadruples the pool
    size and therefore invalidates the assumptions of the "nifty" formula.

    Fix the overflow by using 64bit arithmetics.

    Reported-by: Torsten Hilbrich <torsten.hilbrich@secunet.com>

    This bug is at worst a privileged DoS -- with RANDNET enabled, an admin
    with CAP_SYS_ADMIN feeding large amounts of entropy into the pool at once
    can cause less than expected entropy to be credited (but this doesn't
    affect how much is actually added).  For specific buffer sizes, this
    can result in 0 entropy being credited and end in a situation in which
    the kernel can't recover, causing future reads from /dev/random to stall.

    Many thanks to Torsten and Mathias for the report!

 drivers/char/random.c |    5 +----
 1 files changed, 1 insertions(+), 4 deletions(-)

Signature-tree: 2271bafe567ccf0aa21eea2d00742504b3c21b0a

+++ drivers/char/random.c
@@ -291,9 +291,6 @@
 /*
  * To allow fractional bits to be tracked, the entropy_count field is
  * denominated in units of 1/8th bits.
- *
- * 2*(ENTROPY_SHIFT + log2(poolbits)) must <= 31, or the multiply in
- * credit_entropy_bits() needs to be 64 bits wide.
  */
 #define ENTROPY_SHIFT 3
 #define ENTROPY_BITS(r) ((r)->entropy_count >> ENTROPY_SHIFT)
@@ -646,7 +643,7 @@ retry:
 		/* The +2 corresponds to the /4 in the denominator */
 
 		do {
-			unsigned int anfrac = min(pnfrac, pool_size/2);
+			u64 anfrac = min(pnfrac, pool_size/2);
 			unsigned int add =
 				((pool_size - entropy_count)*anfrac*3) >> s;
